improve billing search endpoint
